Only oad integrations that are bundled by the user.
This is a breaking API change.
For aar users, the behaviour stays the same, but for jar users, they
will have to manually add all the integrations they want to bundle
themselves. if they enable an integration on the server but it is not
bundled, it will just route the requests though our server instead of
using the bundled integration.