Stream batches to URLConnection
This saves memory since we don't have to buffer all payloads. Saves
cycles since we don't need to deserialize the data from disk, we can
just stream it to the connection.

Also adds an abstraction for faking the base url for our API, and
offloading tests that relied on the network to offline.