Embed integrations object into each message
This fixes a bug that would occur when an integration was bundled and
marked with `true`. The `true` would override the bundled `false` in the
integration object of the message, and would be sent server side, as
well as client side.

This updates the code so that we merge the two and save it to disk, at
the cost of increasing our payload sizes.

Also fixes an edge case that could occur when a bundled integration was
enabled after the message got enqueued. THis way the message object
holds the truth about the integration as of the time of creation, rather
than when it gets uploaded.