Don't use cache when forwarding events.
In backup mode, we enable just the Segment integration, but we don't
save it to the disk cache so that on the next run we can still ask for
an update from the server.

This causes the disk cache to return null when running an operation,
throwing a NPE for track events.

The fix is to keep an extra variable just for the project settings in
use, and the cache is only used to save the settings, or retrieve from
it during initialization.