Refactor Options handling
Previously if you sent `All: false`, we'd only use the tracking plan to
figure out if the track event should be sent to integrations.

This updates is so that the Options object takes precedence over all
else. The event tracking plan cannot override options and only fill in
missing pieces.