Correct Options handling
Another fix for handling how options interacts with the tracking plan.
Previously in a102464, the code would always win. There is one case
however in which we want the tracking plan to win in a conflict. When an
event is disabled in the tracking plan (`enabled: false`), we shouldn't
send it to any integrations (including Segment) regardless of the
options provided in code.

All other cases stay the same.