Wait for advertising ID to be populated before enqueing events.
Retrieving the advertising ID from Play Services is asynchronous. This
means events fired before we have the advertising ID may not have this
information available.

This commit makes a few changes to deal with this. First we setup a
CountDownLatch to synchronize between the GetAdvertisindIdTask and the
enqueuing code. This has the potential to block enqueues for a while, so
we implement a 5 second timeout for waiting. The latch is released
either if play services is not available or when the task completes.

Additionally, this makes the code to fire application lifecycle events
async as well, so that we don't block initialization for too long.