Fix Runscope Token escaping when not set.
When folks don't want to set the Runscope token (e.g. contributors), we were incorrectly escaping the keys. This would cause the generated code to be invalid:

```
public static final String RUNSCOPE_TOKEN = ;
```

This causes compilation to fail with the error https://cl.ly/0Z2d101g2y0m.

To fix this, we tweak how the token is rendererd, and always wrap it with quotes. Tweaked the CI config to not wrap the token in quotes in the properties file.

Ref: LIB-221