Set the userId on the analytics executor thread.
Previously we were setting the userId on the thread that `identify` is called (which may be any thread), but reading the userId on a background thread.

Because of this, the userId on events called before identify may have the userId from "newer" events.

```
analytics.identify("foo");
analytics.track("event"); -> may have userId foo
analytics.identify("bar");
```

To fix this, we now also set the userId on the background thread.

Originally referenced in https://github.com/segmentio/analytics-android/issues/522. Note that my original solution proposed we read the userId on the main thread to fix the issue, but the fix implemented here is to read the userId on the background thread instead.

Ref: LIB-120

Closes #522