Retry HTTP 429 Errors
As per the library guidelines, our retry policy should be:

1. Retry network errors (socket timed out, etc.)
2. Retry server errors (HTTP 5xx)
3. Do not retry client errors (HTTP 4xx)
  1. Except HTTP 429.

We weren't handling the HTTP 429 case correctly. This PR updates the retry logic to correctly handle this case. This shouldn't matter too much in practice immediately because we don't serve this status code currently, but is good to have to future proof the library.