replace problematic characters
These characters cause compilation to fail locally on my machine.

```
>>>  ./gradlew clean build

> Configure project :analytics
The ConfigurableReport.setDestination(Object) method has been deprecated and is scheduled to be removed in Gradle 5.0. Please use the method ConfigurableReport.setDestination(File) instead.

> Task :analytics:androidJavadocs
/Users/prateek/dev/src/github.com/segmentio/analytics-android/analytics/src/main/java/com/segment/analytics/internal/Iso8601Utils.java:26: error: unmappable character for encoding ASCII
 * Jackson???s date formatter, pruned to Moshi's needs. Forked from this file:
          ^
/Users/prateek/dev/src/github.com/segmentio/analytics-android/analytics/src/main/java/com/segment/analytics/internal/Iso8601Utils.java:26: error: unmappable character for encoding ASCII
 * Jackson???s date formatter, pruned to Moshi's needs. Forked from this file:
           ^
/Users/prateek/dev/src/github.com/segmentio/analytics-android/analytics/src/main/java/com/segment/analytics/internal/Iso8601Utils.java:26: error: unmappable character for encoding ASCII
 * Jackson???s date formatter, pruned to Moshi's needs. Forked from this file:

```

Replacing these characters fixes the problem locally. I didn't have much luck fixing this in the gradle config directly, and this doesn't fail in CI, so it's likely this is due to a local setup issue, but it's not clear what the issue is.