Use a rule to ignore end to end tests.
Instead of adding an `if (BuildConfig.RUN_E2E_TESTS)` check in each test method, moving this logic into a rule that gets applied for all end to end tests.

The advantage isn't immediately obvious here yet, but I'm adding more tests/cases to this, and this will be useful to avoid copy pasting the code into each method.