Send Slack notifications on failed builds on selected branches
CircleCI 2.0 doesn't support sending notifications for speficic
branches - it notifies on all branches once notifications are turned on.
This is quite noisy.

notify-slack script provides the needed functionality. It should be
called with on_success or on_fail as a build step.

SLACK_WEBHOOK_URL environment variable needs to be set up in CircleCI
with the appropriate Slack webhook url.